+++ src/mml_utils/review/build_afep_excel.py
@@ -26,6 +26,24 @@ def format_table_to_excel(writer, df, name, how='mean'):
             ws.set_column(i, i, max(len(col), 12))
 
 
+def add_diffs(cdf, names):
+    """Add a table comparing results from different vocabularies."""
+    diff_df = None
+    cols = ['cui', 'preferredname']
+    for i, name in enumerate(names):
+        for name2 in names[i + 1:]:
+            res1 = cdf[(cdf[name] == 1) & (cdf[name2] == 0)][cols].copy()
+            res1[f'{name}-vs-{name2}'] = f'{name}_only'
+            res2 = cdf[(cdf[name] == 0) & (cdf[name2] == 1)][cols].copy()
+            res2[f'{name}-vs-{name2}'] = f'{name2}_only'
+            df = pd.concat((res1, res2))
+            if diff_df is None:
+                diff_df = df
+            else:
+                diff_df = pd.merge(diff_df, df, on=cols, how='outer').fillna('NONE')
+    return diff_df
+
+
 @click.command()
 @click.argument('afep_path', type=click.Path(path_type=Path, file_okay=False))
 @click.option('--how', type=str,
@@ -35,7 +53,7 @@ def build_afep_excel(afep_path: Path, how='mean'):
     cdf = None
     writer = pd.ExcelWriter(afep_path / 'afep_summary.xlsx', engine='xlsxwriter')
     names = []
-    for p in afep_path.iterdir():
+    for p in sorted(afep_path.iterdir()):
         if p.is_dir() and 'selected' in p.stem:
             name = p.stem.replace('-selected', '')
             names.append(name)
@@ -54,5 +72,11 @@ def build_afep_excel(afep_path: Path, how='mean'):
     for name in names:
         cdf[name] = cdf[name].apply(int)
     format_table_to_excel(writer, cdf, 'afep_summary', how=how)
+    diff_df = add_diffs(cdf, names)
+    format_table_to_excel(writer, diff_df, 'diffs', how=how)
     writer.close()
     cdf.to_csv(afep_path / 'afep_summary.csv')
+
+
+if __name__ == '__main__':
+    build_afep_excel()